Add files via upload
This is the version immediately after I passed all the autotests and finally finished stage 4. This was definitely a challenge with many brain-wracking parts. One memorable example is where I got the gradients logic reversed since in regular 2d geometry y points up when positive but in a 2d array y points down when positive. Very happy to finally get this done, now time to clean it up.